pyproject.toml
[tool.poetry]
name = "pilligrim"
version = "0.1.0"
description = ""
authors = ["Vas <vas@va.va>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1"
django-cors-headers = "3.13.0"
djangorestframework-simplejwt = "5.2.0"
djangorestframework = "3.13.1"
drf-yasg = "1.21.3"
gunicorn = "20.1.0"
mypy = "0.991"
psycopg2-binary = "2.9.3"
channels = "^4.0.0"
channels-postgres = "^0.0.5"
daphne = "^4.0.0"
djangochannelsrestframework = "^1.1.0"
pip = "^23.0.1"
django-stubs = "^1.15.0"


[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
